Tidied up and bug squishing
Tidies up the AJAX code to be contained in one js file. renamed the
functions to greater reflect their role, also will help with adding more
functions that would have the same name. Fixed a little bug that Kim
Introduced. One stored procedure needed to return a value, changed the
procedure and the code to handle it.